Break out pandas serialization code into separate package for reuse
